requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
PyGithub==2.1.1
aiohttp==3.8.5